# docs/openapi-v2.2.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "剧院死神后端",
    "version": "2.2",
    "description": "数字UID登录、可改昵称、直接注册与管理员账号管理契约。API版本2、玩法规则2.0。"
  },
  "servers": [
    {
      "url": "/api/v2"
    }
  ],
  "security": [
    {
      "accountCookie": []
    }
  ],
  "paths": {
...
    "/auth/register": {
      "parameters": [],
      "post": {
        "summary": "直接注册、分配数字UID并设置会话Cookie",
        "description": "相同requestId、昵称和密码可恢复已完成注册；不同载荷复用ID被拒绝。",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "requestId": {
                    "type": "string"
                  },
                  "nickname": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "requestId",
                  "nickname",
                  "password"
                ],
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthMe"
                }
              }
            }
          },
          "default": {
            "description": "失败。按error.code处理；未知结果不能视为操作已失败。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "parameters": [],
      "post": {
        "summary": "登录（不自动接管房间）",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "uid": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "uid",
                  "password"
                ],
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthMe"
                }
              }
            }
          },
          "default": {
            "description": "失败。按error.code处理；未知结果不能视为操作已失败。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "parameters": [],
      "get": {
        "summary": "读取账号与资料",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthMe"
                }
              }
            }
          },
          "default": {
            "description": "失败。按error.code处理；未知结果不能视为操作已失败。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "parameters": [],
      "post": {
        "summary": "注销当前会话；控制会话注销时离开房间",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {},
                "required": [],
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "loggedOut": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "loggedOut"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "失败。按error.code处理；未知结果不能视为操作已失败。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/change-password": {
      "parameters": [],
      "post": {
        "summary": "修改密码并撤销该账号的全部旧会话",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "currentPassword": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "currentPassword",
                  "password"
                ],
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "changed": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "changed"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "失败。按error.code处理；未知结果不能视为操作已失败。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/rooms/{code}/second-screen/invitations": {
      "parameters": [
        {
          "name": "code",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "post": {
        "summary": "本人签发一次性第二屏邀请",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "requestId": {
                    "$ref": "#/components/schemas/RequestId"
                  },
                  "gameId": {
                    "type": "string"
                  }
                },
                "required": [
                  "requestId",
                  "gameId"
                ],
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string"
                    },
                    "expiresAt": {
                      "type": "integer"
                    },
                    "gameId": {
                      "type": "string"
                    },
                    "subjectPlayerId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "token",
                    "expiresAt",
                    "gameId",
                    "subjectPlayerId"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "409": {
            "description": "状态或幂等冲突，包括 stale_game、request_id_reused。相同ID重试保持完整载荷；命令入队后的拒绝另见200回执。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "description": "失败。按error.code处理；未知结果不能视为操作已失败。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/rooms/{code}/second-screen/redeem": {
      "parameters": [
        {
          "name": "code",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "post": {
        "summary": "原子进入/升级私人第二屏",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "requestId": {
                    "$ref": "#/components/schemas/RequestId"
                  },
                  "gameId": {
                    "type": "string"
                  },
                  "token": {
                    "type": "string"
                  }
                },
                "required": [
                  "requestId",
                  "gameId",
                  "token"
                ],
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "memberId": {
                      "type": "string"
                    },
                    "gameId": {
                      "type": "string"
                    },
                    "subjectPlayerId": {
                      "type": "string"
                    },
                    "kind": {
                      "type": "string",
                      "enum": [
                        "private_spectator"
                      ]
                    }
                  },
                  "required": [
                    "memberId",
                    "gameId",
                    "subjectPlayerId",
                    "kind"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "409": {
            "description": "状态或幂等冲突，包括 stale_game、request_id_reused。相同ID重试保持完整载荷；命令入队后的拒绝另见200回执。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "description": "失败。按error.code处理；未知结果不能视为操作已失败。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/rooms/{code}/second-screen/revoke": {
      "parameters": [
        {
          "name": "code",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "post": {
        "summary": "撤销本人签发的私有授权和未用邀请",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "requestId": {
                    "$ref": "#/components/schemas/RequestId"
                  },
                  "gameId": {
                    "type": "string"
                  }
                },
                "required": [
                  "requestId",
                  "gameId"
                ],
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "revoked": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "revoked"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "409": {
            "description": "状态或幂等冲突，包括 stale_game、request_id_reused。相同ID重试保持完整载荷；命令入队后的拒绝另见200回执。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "description": "失败。按error.code处理；未知结果不能视为操作已失败。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "livekitSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "LiveKit服务端签名；SDK验证原始body。"
      },
      "accountCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "td_account_v2",
        "description": "HttpOnly会话，浏览器自动携带。隔离开发3003使用td_account_contract_21，按服务配置区分。"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
      "RequestId": {
        "type": "string",
        "minLength": 1,
        "maxLength": 80,
        "description": "一次意图的稳定编号，建议UUID。同ID重试必须保持完全相同载荷。"
      },
...
      "AuthMe": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "uid": {
            "type": "string"
          },
          "nickname": {
            "type": "string"
          },
          "avatarUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "profileVersion": {
            "type": "integer"
          },
          "expiresAt": {
            "type": "integer"
          }
        },
        "required": [
          "userId",
          "uid",
          "nickname",
          "avatarUrl",
          "profileVersion",
          "expiresAt"
        ],
        "additionalProperties": false
      },
...
    }
  },
...
}